Initial commit of the lexer
This is the initial commit for the Vase compiler project.
Currently, only the lexer has been 'completed'.

I've tried to keep the code fairly clean and somewhat rusty.
I've also documented a lot of the project and have tried to
keep tabs on where things aren't finished or where
improvements can be made. TODO, BUG, and OPTIMIZATION comments
are used to mark these areas.

Currently, most code in vase::c is *not* documented. I've
done this because the code is in flux and is likely to
change.

Right now, the goal moving forward will be to get the parser
functional.