Add license information
I forgot to add the license in the initial commit. All the code in the previous
commit as well as in this repository will be licensed under GPLv3 (or later)
for now.

I wish to retain the ability to re-license the project. In case I decide to
change the license later on.