Add AtomicArc
AtomicArc will be my replacement for ArcSwapOption. It has only the features I
need and is faster.

I added corresponding benchmarks to compare AtomicArc with AtomicBox and
ArcSwapOption. AtomicArc is only a little slower than AtomicBox (expected
considering the reference count). This means AtomicArc is a magnitude faster
than ArcSwapOption. It's also nice because I can remove a dependency.