Fix #15 - StringCache confuses values
StringCache was confusing values because I replaced a 'swap with none' operation
for a 'swap if none' operation during the ArcSwapOption -> AtomicArc transition.
If I had ran the tests, I would have realized this much earlier on.

This poses a problem, AtomicArc's safety hinges on not have any operation that
changes a non-none value. As such, I was left with three options:

1. Add an unsafe swap_to_none function on AtomicArc. This would then require an
unsafe usage in string_cache.rs. It would still technically be safe, but I am
trying to minimize unsafe code.

2. Separate an end-node's value from a value-node's value. Under rare
circumstances, this means a node can be both a end-node and a value-node.
This approach is 100% safe.

3. Add the back the arc_swap dependency just for this once case. Also 100% safe.

Option #3 was out of the picture for me since ArcSwapOption was quite slow to
both read and write.

I didn't like option #1 since it required more unsafe operations in StringCache.

So I went with option #2 because it's only downside is it required another
field. If I went through and optimized the node-sizes, this is probably
negligible. Since option #2 doesn't require an atomic decrement and nodes
can end up as both an end-node and a value-node, it may end up being more
performant.

Working on this also exposed that my chain-node technique was mildly incorrect.
When it came to strings that different somewhere in the middle, it would
actually place the end value right in the way of the cache-request. The code
could obviously handle this, but it was an inefficiency.