Combine all keywords into one token
Due to the fact that macros can be defined using keywords, I need to know the
unique ID of any keyword. The easiest way to do this is to combine all keywords
into one token and have that token also keep the unique ID. The reason I don't
store the CachedString is because that would create a lot of extra
atomic-increments. We can always get the string of the keyword through a
member function.