Add enum with properties macro
While designing the error collection system, I came across a problem:
each variant of the error enum is going to need separate severities
and error codes. Given that I expect the error enum(s) to have 100+
variants, this would be unmanagable.

I could have defined a trait and then shoved enums into an Arc, except
that defeats the purpose of the enum in the first place. I want errors
to be an enum so that other code could figure out exactly which error
was hit (without relying on magic codes). Sides, unless the error codes
were auto-generated (which has its own problems), error codes would
still not be possible with a trait.

About the only way this could be done without a macro is to use a large
match statement. With 100+ different values, that would be unmanagable.

`enum_with_properties! {}` solves this problem for me. I can define the
property functions I want and then I can attribute each variant with the
values. I personally prefer this solution since it keeps the value next
to the variant.