Add lexer error handling
Lexer errors are now handled using special tokens. The errors themselves
are stored in the token stack to allow easy querying if and what errors
occured.

This also includes fatal errors (IO and UTF-8 errors). These are shoved
into a token stack so that they have the same behavior as other
errors.