Make tokens use the SourceLocation utility
Tokens have, for quite some time, kept their location in 2-3 fields.
However, other code needs to be able to interact with these
locations on their own. This is why SourceLocation was created.

This commit converts all current code over to SourceLocation
and also performs a few necessary changes. I will sum up
everything:

* FileReader no longer tracks length since the `through` method
on a source location can provide this functionality. This has
also allowed CharLocation to be non-public.
* The lexer was 'simplified' by creating a state object. This
prevents having to pass the token stack around to every function.
It also allows errors to be given a location.
* Errors now have a proper location on their tokens.
* Joined tokens now have a location that points to the joiner.