Make enum_with_properties more resilient
enum_with_properties was quite fragile because it had to figure out
when the property functions ended and the enum started. For
example, it's not been possible to annotate property functions
*at all* because the code would think an enum is coming next.

By requiring the property functions be inside an impl-block, it
removes this problem. It also allows the impl block to be
implementing a trait (which may come in handy in the future).