Add number parsing
I looked for a while for a library that could do this for me,
but none fully fit the bill. Here's what I needed:

1. The ability to parse hexadecimal floating point numbers
  - This eliminates the standard library.
2. To have *some* value even if an overflow occured.
  - This eliminates most other libraries.
3. To be able to detect over-precise floating point numbers.
   - This eliminates every library I looked at.

Because of this, I had to implement the parsing myself.
Integer parsing isn't too hard. Real-number parsing is
known to be hard. It is *very* likely that the parser
shown here produces incorrectly-rounded results.

My methods here may not be the fastest, but they should hopefully
get the job done for now.