Report traveler errors to a callback
Figuring out how to handle errors/warnings in the traveler has been
quite the task. If I attempted to just return an error token, that would
limit recover-ability and potentially be a head-ache in the parser.

The only alternative that I could think of, would be to give traveler
a callback to call when an error occurs. So far, this strategy has
worked fairly-okay. I went with a templated callback this time
as it allows a future parser to handle traveler errors.