Make lexer callback generic
By having the include callback be generic, it makes the lexer
far more flexible. The only downside is it destroyed what
limit remained of type-interference. Type-interence with
callbacks seems to be more limited as I've ran into
non-compatible types multiple times. Regardless of that
limitation, I do think this is an improvement.

For this change to be reasonably done, I've also enabled
trait aliases. These allow the repetition of trait bounds
to be less repetitive (not repeating the arguments
everywhere).