Allow clippy::similar_names
clippy::similar_names is a warning with good intentions. However, I
don't think it's a good fit for this project. There are multiple cases
where the similar names are intentional and for the better
(LParen vs RParen). In fact, the warning has yet to prevent me from
writing bad code. All it has done so far is warn on code I wouldn't
have any other way.

Since this is a warning I enabled, that enable has been removed.