Split WChar into WChar16 and WChar32
One thing I recognized while working on #if is that wchar
characters and strings can be 16-bit or 32-bit depending on
the platform. While in theory, I could pass the environment
*everywhere* to keep track of this mess, I decided the better
solution is to split WChar into WChar16 and WChar32.

When compiling for a target that has a 16-bit wchar, WChar16
will be used (WChar32 shouldn't even show up).
Likewise for 32-bit wchar and WChar32.