Add mathematical module and utilities
While working on #if/#elif I've been running into the problem of
either copy-pasting code for unsigned/signed or defining *yet*
another trait.

Because of this conundrum, I'm just going to define just 3 traits
to handle numbers: Number, Integer, and Real. Number is the
operations that can be performed on both integers and reals.
Integer and Real are fairly obvious.

These traits will allow some code to be less-duplicated.

I've also added the Sign enum. The expressions in #if/#elif are
done entirely in i64 and u64. This enum is a small (non-language
dependant) utility around that concept.

I am planning on moving num_parsing to the math folder in
a future commit.