Fix #17 - #if and #elif support
This implements support for #if and #elif for everything but
character literals (which can easily be done later). IfParser
takes the traveler and walks through the expression tokens
to build a simplified AST. That AST is then handed to IfEvaluator.
IfEvaluator then walks the AST to figure out the value of the
expression.

-----------  Side Note  ---------------
Working on error reporting and #if/#elif has been quite the journey.
Each time I thought I was getting close to implementing #if/#elif,
I'd realize I needed to implement *another* utility to support
thing-X in a sensible manner.

Working on these issues has also made me develop my first
rust macros. If it wasn't for the amount of time it can take to
implement a macro, I'd probably be using them more. The
two that I've made so far, enum_with_properties and create_intos
has probably saved me time at this point.