Add CodedError utility
Almost all the errors in the c-module have a severity property
and a code property. The goal of this trait is to describe that
behavior with a trait. Further error traits may be added in the
future.

I've decided to split the code property into the code's number
and the code's prefix. This was done to potentially allow errors
to redefine only one part of another error's code.