Rename and move c::ErrorScope to error::Unwind
ErrorScope was always a terrible name for the concept I wanted to
capture. I think 'Unwind' is a better name because it better describes
what it represents: how far to unwind the function call stack.

This also produced a beautiful name for the result alias: MayUnwind.
Either the function completed successfully or it was unwound due to
an error.

I've also moved Unwind/MayUnwind to the error module since unwinding
occurs due to errors.