Abbreviate location as loc everywhere
Generally I avoid using abbreviations, however,  there are times where
an abbreviation has it's clearer. While working on #if/#elif and the
AST, I found abbreviating Expression as Expr reduced noise. I then
used expr everywhere for expression.

I've found using loc for location has a similar effect. Location is long
and it tends to crowd variable names. This is why, when designing
the AST, I used things like if_loc and op_loc. Of course, an
abbreviation needs to be used *consistently*. As such, I've renamed
everything with location in it to use loc.