Remove excess traveler error types
Errors like DefineMissingId can be folded into DefineExpectedId. I've
already planned on the same error having potentially different
messages. The reason for this is to provide better context depending
on the scenario (EX: `#define 0` can mention that you can't define
numbers where as `#define {` can't say anything specific).

In the case of errors like DefineMissingId, the text is just a special
case of UndefExpectedId. This case occurs when the token it contains
is a PreEnd token.