Add dependencies and index to travel state
One thing that I realized would be nice for the state to store is all
the file ids that have been used so far. This could allow dependency
analysis later on.

I've also added what I call a TravelIndex. One of the realizations I
had, was I had no plan for *semantic* errors. Lexer and parser
errors I can rely on the saved state. It's unreasonable to store
so many states, so I've decided to save a travel index. This index
could be used later on to bring a traveler to the same spot.

This does sadly mean error reporting may be a bit slower than
necessary. However, I have ideas on how to optimize it when
the time comes.