Optimize CharExt::decode_utf8
While testing the current performance of the compiler, I've found that
most of the application's time is in the lexer. I expected the traveler
to take the most time, so this is surprising.

Looking at the profiling data, there are a few red flags:
* CharExt::decode_utf8 was taking more time than I would expect.
* Lexing identifiers takes about half of the lexer's time (!).
  * Sadly, half of this time is in the string cache.
* It seems like a lot of blocking is occuring when reading files.

Out of these, CharExt::decode_utf8 was obvious on how to solve:
remove the leading_ones call. Most of the decoding time was
spend in that function. Removing it for a simple match statement
fixed the performance problem. In fact, the match statement is
more than *twice* as performant as leading_ones.