Fix #23 - Add internal headers
This adds the standard headers that are not included in /usr/include.
These headers tend to be in a compiler-specific directory. However,
I can't just rely on those directories because the directory paths
vary based on the version of the compiler. As such, I needed to package
these headers with the application.

For now, I've done this by adding an include directory that *must* be
in the same directory as executable.