Add ErrorReceiver utility
I've found that the generics over FnMut are a little restrictive
currently. Right now, one can not impl a FnMut trait on an
object (in stable). However, what I can do is define a trait and
then implement that trait for FnMut objects.

I've also implemented the trait for Arc-RefCells of the trait.
I may end up using this in the parser since the parser has
to report both traveler and parser errors to a single callback.