Add a lifetime parameter to the traveler
Sadly, to get the error-forwarding to work in the parser,
I've had to add a fake lifetime to the traveler. While the code
diff is small, this was definitely a battle with the borrow-checker
that wasted a lot of time.

The sad part is, this contortion wouldn't be necessary if we could just
add lifetime annotations to the fields of a struct.