Add AssignOp
I was originally going to re-use BinaryOp for assignment expressions but
I've found this to be unwieldy (since there is the normal assignment
operator).

As such, I've created AssignOp to represent the different types of
assignments. I've also decided to use 'assign' instead of 'assignment'
everywhere. It communicates the same idea but is more concise.