Add Conversions trait; a generic try_into
I've had a few problems with Rust's default try_into: there is no way to
specify the type. Normally the type-interface can handle this, but there
have been a few times where it can't.

The Conversions trait is an extension trait that provides a generic
version of .into (named .into_type) and .try_into (named .try_into).
Since the Into trait is imported by default, the function has to have
a different name. I decided to *not* have a different name for
.try_into since no collision occurs (unless TryInto is imported).
The Conversions trait can just be a replacement for TryInto in
most code.