Remove TokenKind::is_binary_op
I've decided to remove is_binary_op since I would also need an
equivalent for assign operators. Rather than detecting the
tokens in a match, I've decided to use an if-let with an else
clause that contains the rest of the match.

If the if_let_guard feature was supported in rust-analyzer, I'd use
it.