Make file reading and multithreading optional
Due to time constraints, I'm planning on re-using the code I've made
here for an all-at-once syntax highlighter (so that I have something to
show).

Part of this plan involves creating a webpage to do this. As such,
vase (as a library) needs to be able to compile to web-assembly.

Thankfully, all that is required to make webassembly work is
to make file-reading and multithreading optional.