Improve atomic code
This helps to standardize the interfaces of both AtomicArc and AtomicBox
(which previously differed for no good reason). Both AtomicArc and
AtomicBox now have the same set of operations (excluding Arc-specific
ones). The names for these operations were selected to mirror that of
other atomic/optional types in Rust.

The documentation of AtomicArc and AtomicBox are now fairly similar
since they have similar interfaces.

I've removed all the ordering parameters to AtomicArc/AtomicBox because
the orderings shouldn't differ between use-cases. AtomicArc/AtomicBox
have a limited set of operations. I think the atomic orderings can be
defined based on those orderings.

OnceArray now operates on AtomicArcs rather than AtomicBoxes. This was
done because the only use-case for OnceArray in vase, is to store an
array of Arcs. In the future, it is possible to have OnceArray able to
do both (through the use of a trait). This wasn't done since it isn't
needed.

Here's a fairly complete list of changes:

* Both AtomicArc and AtomicBox
  * `get_or_else` function to get the current value or create one
  * `load_or_else` as an atomic cousin to `get_or_else`
  * `set` now takes an Option (so None can now be set)
  * implement Default rather than having an empty function.
  * All orderings have been removed.
* AtomicArc
  * Now has a `from_raw` function like AtomicBox.
* AtomicBox
  * `try_set_if_none` rather than `try_set_if_null`
* OnceArray
  * Uses AtomicArc rather than AtomicBox for the values
  * Can set an index using mutability (rather than the limited atomic
     operation).