Remove warnings caused by disabled feature
When the file-reading feature is disabled, two warnings would occur in
the project.

One occured in the lexer due to an unused File import (since the file
that used the import was disabled).

The other occured in CompileSettings::new due to an unnecessarily
mutable variable (due to the muting part of the code being
disabled).