Use cached strings instead of numbers for macros
Macros have been using a 'unique id' system for some time. The
uniqueness of this number depends on the cached string. This
system also required an extra field in the keyword token kind.

I've decided to use the cached strings instead of the magical
numbers. While this is most likely slower, I believe it makes the
code clearer.