Add RedeclMap utility
One problem I had while designing the parser and AST is handling
redeclarations in a sane matter. Essentially, I needed a collection
that could handle:
1. Values that didn't have a key (for inner structs)
2. Redeclarations of any key with another value
  - This is for both forward declarations as well as mistaken
  double-declarations.
3. Redeclaring a key doesn't destroy the old value
4. A way to retrieve old values

Because of this, I've created RedeclMap. It can handled values without
a key (#1) and any key can be redeclared without losing old values
(#2 and #3). The indexes that are returned when values are added
provide a way to access the value despite redeclarations (#4).

Part of it's design decision was to go with a pair of indexes to
locate specific values. This makes the index cheap to copy
and requires no atomics (unlike CachedStrings).